+++ src/Market/ChinaMarket.cpp
@@ -881,7 +881,7 @@ void CChinaMarket::DeleteDayLine(chrono::local_days lDate) const {
 	auto db = gl_dbStockMarket.get();
 	auto tx = sqlpp::start_transaction(db);
 
-	db(sqlpp::delete_from(t).where(t.Date == static_cast<int>(toFormattedDate(lDate))));
+	db(sqlpp::delete_from(t).where(t.Date == toFormattedDate(lDate)));
 	tx.commit();
 }
 
@@ -903,13 +903,13 @@ void CChinaMarket::UpdateOptionDB() {
 		auto result = db(select(all_of(t)).from(t));
 		if (result.size() == 0) {
 			db(sqlpp::insert_into(t).set(
-				t.LastLoginDate = static_cast<int>(toFormattedDate(GetMarketDate())),
+				t.LastLoginDate = toFormattedDate(GetMarketDate()),
 				t.LastLoginTime = toFormattedTime(GetMarketTime())
 			));
 		}
 		else {
 			db(update(t).set(
-				t.LastLoginDate = static_cast<int>(toFormattedDate(GetMarketDate())),
+				t.LastLoginDate = toFormattedDate(GetMarketDate()),
 				t.LastLoginTime = toFormattedTime(GetMarketTime())
 			));
 		}
